+++ src/policy/index.ts
@@ -3,3 +3,4 @@ export * from "./detect-framework";
 export * from "./file-size";
 export * from "./patterns";
 export * from "./evaluate";
+export * from "./apex";